+++ brightwind/analyse/plot.py
@@ -318,7 +318,7 @@ def plot_monthly_means(data, coverage=None, ylbl=''):
 
 def _timeseries_subplot(x, y, x_label=None, y_label=None, x_limits=None, y_limits=None, x_tick_label_angle=25,
                         line_marker_types=None, line_colors=None, subplot_title=None,
-                        legend=True, ax=None):
+                        legend=True, ax=None, external_legend=False, show_grid=False):
     """
     Plots a timeseries subplot where x is the time axis.
 
@@ -360,6 +360,10 @@ def _timeseries_subplot(x, y, x_label=None, y_label=None, x_limits=None, y_limit
     :param ax:                      Subplot axes to which assign the subplot to in a plot. If None then a single plot is
                                     generated
     :type ax:                       matplotlib.axes._subplots.AxesSubplot or None
+    :param external_legend:         Flag for option to return legend outside and above the plot area, default False
+    :type external_legend:          bool
+    :param show_grid:               Flag for option to add a grid to the plot area, default False
+    :type show_grid:                bool
     :return:                        A timeseries subplot
     :rtype:                         matplotlib.axes._subplots.AxesSubplot
 
@@ -463,7 +467,12 @@ def _timeseries_subplot(x, y, x_label=None, y_label=None, x_limits=None, y_limit
     ax.tick_params(axis="x", rotation=x_tick_label_angle)
 
     if legend:
-        ax.legend()
+        legend_kwargs = {
+            'bbox_to_anchor':(0, 1.02, 1, 0.2), 'loc':'lower left', 'mode':'expand', 'ncol':5 
+            } if external_legend else {}
+        ax.legend(**legend_kwargs)
+    if show_grid:
+        ax.grid()
 
     if subplot_title is not None:
         ax.set_title(subplot_title, fontsize=mpl.rcParams['axes.labelsize'])